Fix flash of "No results found" message
If the user enters search text for a sub-search
and then escapes out of the quick-switcher, when
re-opening the quick-switcher the previously entered
search text was briefly being used to search for
results in the main menu.  Clearing out the search text
before hiding/showing the quick-switcher will prevent
the main search from being queried with the previous
value.