Switch docs from using dist files to Jekyll
The scss file cannot be directly used in browser, so I was utilizing the
CSS file that was being compiled into the `dist` directory. I am wanting
to get rid of the `dist` directory (in favor of uploading them to GitHub
releases) and start using Jekyll for documentation. Using Jekyll to
handle the scss is a good start to both goals.