Support initializing multiple quick switchers
In order to allow multiple quick switchers to be initialized, I need
to make the quick switcher's visibility decoupled from global state.
As such, using the noscroll class on the body element to decide if the
quick switcher should be visible no longer works. The quick switcher's
visibility needs to be manipulated directly.

The quick switcher now uses .show() and .hide() to control the quick
switcher's visibility

Issue #73