Add the ability to prevent search text clearing
Sometimes we want a subsearch to be always-available in the
quick-switcher so that a search term can be typed and the search term
can be transferred to the subsearch's filter. For example, if there is
a subsearch for colors, we may want the user to be able to type 'green'
on the main search and have an item such as "Search for 'green' colors"
appear. On selection of the subsearch, the subsearch would become active
and 'green' would automatically be searched for within the subsearch.